# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PhoAmThuc API",
    "description": "Auto-generated OpenAPI spec from Express routes.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "description": "Create a new account with email and password.",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "fullName": {
                    "example": "any"
                  },
                  "deviceId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [],
        "operationId": "post_api_v1_auth_register",
        "summary": "Register account"
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Authenticate user and return auth tokens.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "deviceId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [],
        "operationId": "post_api_v1_auth_login",
        "summary": "Login account"
      }
    },
    "/api/v1/auth/payment/claim": {
      "post": {
        "description": "Redeem access code for authenticated user.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "post_api_v1_auth_payment_claim",
        "summary": "Redeem claim code"
      }
    },
    "/api/v1/auth/token-refresh": {
      "post": {
        "description": "Refresh auth session by refresh token.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthRefreshResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "refreshToken"
                ],
                "properties": {
                  "refreshToken": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [],
        "operationId": "post_api_v1_auth_token_refresh",
        "summary": "Refresh access token"
      }
    },
    "/api/v1/auth/payment/initiate": {
      "post": {
        "description": "Create payment transaction for authenticated user.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentInitiateResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount"
                ],
                "properties": {
                  "paymentMethod": {
                    "type": "string",
                    "enum": [
                      "vnpay",
                      "momo"
                    ]
                  },
                  "provider": {
                    "type": "string",
                    "enum": [
                      "vnpay",
                      "momo"
                    ]
                  },
                  "amount": {
                    "type": "number"
                  },
                  "currency": {
                    "type": "string",
                    "default": "VND"
                  },
                  "deviceId": {
                    "type": "string"
                  },
                  "returnUrl": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "post_api_v1_auth_payment_initiate",
        "summary": "Initiate payment"
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Revoke current access token from header or request body.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "post_api_v1_auth_logout",
        "summary": "Logout session"
      }
    },
    "/api/v1/auth/payment/callback": {
      "post": {
        "description": "Validate callback signature and finalize payment status.",
        "parameters": [
          {
            "name": "x-idempotency-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-callback-signature",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-callback-timestamp",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentCallbackResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "orderId": {
                    "type": "string"
                  },
                  "transactionId": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "success",
                      "failed",
                      "cancelled",
                      "succeeded",
                      "fail",
                      "canceled"
                    ]
                  },
                  "provider": {
                    "type": "string",
                    "enum": [
                      "vnpay",
                      "momo"
                    ]
                  },
                  "gatewayPayload": {
                    "type": "object"
                  },
                  "deviceId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [],
        "operationId": "post_api_v1_auth_payment_callback",
        "summary": "Handle payment callback"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-api-key"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "Validation failed"
              }
            }
          }
        }
      },
...
      "AuthSessionResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "number"
          },
          "user": {
            "type": "object"
          }
        }
      },
      "AuthRefreshResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "number"
          }
        }
      },
      "PaymentInitiateResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "transactionId": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "paymentUrl": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "LogoutResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "PaymentCallbackResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "expiresIn": {
            "type": "number"
          },
          "deviceId": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}